Feat: replacing config and updating its test.
