+++ src/config.spec.ts
@@ -1,39 +1,45 @@
 import { describe, expect, test, vi } from "vitest";
 import { z } from "zod";
-import { createConfig } from "./config";
+import { Config, createConfig } from "./config";
 import { AbstractLogger } from "./logger";
 
 describe("Config", () => {
   describe("createConfig()", () => {
-    test("should return its argument", () => {
-      expect(
-        createConfig({
-          emission: {
-            "/": {},
-            test: {},
-          },
-          timeout: 2000,
-          logger: { debug: vi.fn() } as unknown as AbstractLogger,
-        }),
-      ).toEqual({
-        emission: { "/": {}, test: {} },
+    test("should create the class instance from the definition", () => {
+      const config = createConfig({
+        namespaces: {
+          "/": { emission: {}, hooks: {} },
+          test: { emission: {}, hooks: {} },
+        },
         timeout: 2000,
-        logger: { debug: expect.any(Function) },
+        logger: { debug: vi.fn() } as unknown as AbstractLogger,
+      });
+      expect(config).toBeInstanceOf(Config);
+      expect(config.logger).toEqual({ debug: expect.any(Function) });
+      expect(config.timeout).toBe(2000);
+      expect(config.namespaces).toEqual({
+        "/": { emission: {}, hooks: {} },
+        test: { emission: {}, hooks: {} },
       });
     });
 
-    test("should ensure namespaces", () => {
+    test("should set defaults and provide namespace augmentation method", () => {
+      const base = createConfig({});
+      expect(base).toBeInstanceOf(Config);
+      expect(base.logger).toEqual(console);
+      expect(base.timeout).toBe(2000);
+      expect(base.namespaces).toEqual({});
       const schema = z.tuple([]);
-      expect(
-        createConfig({
+      const config = base.addNamespace({
+        path: "/",
+        emission: { test: { schema } },
+      });
+      expect(config).toBeInstanceOf(Config);
+      expect(config.namespaces).toEqual({
+        "/": {
           emission: { test: { schema } },
-          timeout: 2000,
-          logger: { debug: vi.fn() } as unknown as AbstractLogger,
-        }),
-      ).toEqual({
-        emission: { "/": { test: { schema } } },
-        timeout: 2000,
-        logger: { debug: expect.any(Function) },
+          hooks: {},
+        },
       });
     });
   });
